Константы в отдельный файл
