+++ js/api.js
@@ -1,19 +1,4 @@
-const BASE_URL = 'https://31.javascript.htmlacademy.pro/kekstagram';
-
-const Route = {
-  GET_DATA: '/data',
-  SEND_DATA: '/',
-};
-
-const Method = {
-  GET: 'GET',
-  POST: 'POST',
-};
-
-const ErrorText = {
-  GET_DATA: 'Не удалось загрузить данные. Попробуйте обновить страницу',
-  SEND_DATA: 'Не удалось отправить форму. Попробуйте ещё раз',
-};
+import { BASE_URL, Route, Method, ErrorText } from './consts.js';
 
 const load = (route, errorText = null, method = Method.GET, body = null) =>
   fetch(`${BASE_URL}${route}`, {method, body})
